Restructured code under src.
