+++ src/enums/tiled-map-format.enum.ts
@@ -1,0 +1,13 @@
+export enum TiledMapFormat {
+   
+      /**
+       * TMX map layer format
+       * @unsupported
+       */
+      TMX,
+   
+      /**
+       * JSON map layer format
+       */
+      JSON
+   }
