+++ src/interfaces/itiled-map-terrain.interface.ts
@@ -1,0 +1,4 @@
+export interface ITiledMapTerrain {
+   name: string;
+   tile: number;
+}
